+++ tests/loggers/test_seq2seq.py
@@ -180,3 +180,77 @@ def test_log_model_outputs(
             )
 
         assert perplexity > 0
+
+
+def test_model_logger_remove_padding() -> None:
+    """Test _remove_padding and _retrieve_sample_labels
+
+    Ensure that _remove_padding removes the correct tokens for each
+    sample based on the `sample_labels` and the tokenzier padding direction.
+    """
+    tokenized_labels = [
+        np.arange(10).tolist(),
+        np.arange(18).tolist(),
+        np.arange(20).tolist(),
+        np.arange(4).tolist(),
+    ]
+
+    config = mock.MagicMock()
+    config.id_to_tokens = {}
+    config.id_to_tokens["training"] = dict(zip(list(range(4)), tokenized_labels))
+    mock_tokenizer = mock.MagicMock()
+    # First test removing from right padding
+    mock_tokenizer.padding_side = "right"
+    config.tokenizer = mock_tokenizer
+
+    batch_size = 4
+    max_seq_len = 20
+    vocab_size = 100
+
+    logprobs = np.random.rand(batch_size, max_seq_len, vocab_size)
+    # Set pad tokens on the right with -1
+    for idx, token_labels in enumerate(tokenized_labels):
+        logprobs[idx, len(token_labels) :] = -1
+    # Create the top indices just using logits
+    top_indices = logprobs[:, :, 5]
+
+    # Note we don't differentiate between logits and logprobs for this test
+    log_data = dict(
+        ids=list(range(batch_size)),
+        logits=logprobs,
+        split="training",
+        epoch=0,
+    )
+    logger = Seq2SeqModelLogger(**log_data)
+    logger.logger_config = config
+    for sample_id, (sample_logprobs, sample_top_indices) in enumerate(
+        zip(logprobs, top_indices)
+    ):
+        sample_labels = logger._retrieve_sample_labels(sample_id)
+        # Test the retrieve samples method
+        assert np.allclose(sample_labels, tokenized_labels[sample_id])
+
+        no_pad_logprobs, no_pad_top_indices = logger._remove_padding(
+            sample_labels, sample_logprobs, sample_top_indices
+        )
+        assert len(np.where(no_pad_logprobs == -1)[0]) == 0
+        assert len(np.where(no_pad_top_indices == -1)[0]) == 0
+
+    # Test padding on the 'left'
+    logger.logger_config.tokenizer.padding_side = "left"
+    logprobs = np.random.rand(batch_size, max_seq_len, vocab_size)
+    # Set pad tokens on the left with -1
+    for idx, token_labels in enumerate(tokenized_labels):
+        logprobs[idx, : -len(token_labels)] = -1
+    # Create the top indices just using logits
+    top_indices = logprobs[:, :, 5]
+
+    for sample_id, (sample_logprobs, sample_top_indices) in enumerate(
+        zip(logprobs, top_indices)
+    ):
+        sample_labels = logger._retrieve_sample_labels(sample_id)
+        no_pad_logprobs, no_pad_top_indices = logger._remove_padding(
+            sample_labels, sample_logprobs, sample_top_indices
+        )
+        assert len(np.where(no_pad_logprobs == -1)[0]) == 0
+        assert len(np.where(no_pad_top_indices == -1)[0]) == 0
